requirements.txt
art==5.8
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
idna==3.4
lxml==4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
Pillow==9.4.0
Pygments==2.14.0
python-docx==0.8.11
python-dotenv==1.0.0
python-pptx==0.6.21
requests==2.28.2
rich==13.3.2
soupsieve==2.4
tqdm==4.64.1
typer==0.7.0
urllib3==1.26.14
XlsxWriter==3.0.8
